chore(stencil): pin stencil version to 3.2.2
Stencil release after 3.2.2 break our unit tests. This seems to be related to the upgrade to
typescript 5, but so far it's been difficult to debug. Rather than burning to much time on it
now, we'll pin the version and see if we can handle the test/build issues in a future change.